Fixed duplicate seren spirit detection
When the chatbox has an overlay on top (such as right click) an event is fired again for the chat which causes duplicate seren spirit notifications.
We just store the last detected seren spirit and double check they are not the same on next chat detection.
in instances where the user has timestamps turned off and has the same two seren spirits in a row this will more than likely not detect the 2nd seren spirit drop